+++ changelog.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Added `--force-relay` and `--force-direct` CLI flags that control the transit connection
+	- The feature is also exposed in the API
+
 ## Version 0.3.0
 
 - Added experimental port forwarding feature